requirements.txt
flask>=3.0.0
flask-wtf>=1.2.0
werkzeug>=3.0.0
requests[socks]>=2.25.0
APScheduler>=3.10.0
croniter>=1.3.0
bcrypt>=4.0.0
cryptography>=41.0.0
pystray>=0.19.5
Pillow>=10.0.0
